+++ lib/src/widgets/custom_snackbar.dart
@@ -6,10 +6,9 @@ import 'package:flutter/material.dart';
 /// An enumeration that defines the visual style (or "vibe") for the SnackBar.
 ///
 /// This enum is used to determine the color scheme and icon for the custom SnackBar.
-/// 
+///
 enum SnackBarVibe {
   error,
-  neutral,
   success,
   warning,
   none,
@@ -40,52 +39,40 @@ class CustomSnackBar {
   /// - [shape]: Optional. Defines the shape of the SnackBar. Defaults to a [RoundedRectangleBorder] with a stadium border.
   /// - [duration]: Optional. Specifies the duration for which the SnackBar is visible.
   ///   Defaults to 2000 milliseconds.
-  static void showSnackBar(
-      BuildContext context, {
-        required String content,
-        SnackBarVibe vibe = SnackBarVibe.neutral,
-        Color? backgroundColor,
-        TextStyle? textStyle,
-        Widget? icon,
-        Widget? contentWidget,
-        DismissDirection? dismissDirection,
-        ShapeBorder? shape,
-        Duration? duration,
-        EdgeInsets? margin
-      }) {
-
-
+  static void showSnackBar(BuildContext context,
+      {required String content,
+      SnackBarVibe vibe = SnackBarVibe.none,
+      Color? backgroundColor,
+      TextStyle? textStyle,
+      Widget? icon,
+      Widget? contentWidget,
+      DismissDirection? dismissDirection,
+      ShapeBorder? shape,
+      Duration? duration,
+      EdgeInsets? margin,
+      bool usePrimaryColor = false,
+      bool? showCloseIcon,
+      EdgeInsets? padding,
+      double? width,
+      Color? borderColor}) {
     final Map<SnackBarVibe, Color> mainColors = {
-      SnackBarVibe.none: Colors.grey.shade800,
+      SnackBarVibe.none: Colors.blueGrey.shade800,
       SnackBarVibe.info: Colors.lightBlue.shade800,
-      SnackBarVibe.error: Colors.red.shade800,
-      SnackBarVibe.neutral: Colors.blueGrey.shade800,
-      SnackBarVibe.success: Colors.green.shade800,
-      // Use the provided pumpkin color for warning
+      SnackBarVibe.error: const Color(0xfff30d0d),
+      SnackBarVibe.success: const Color(0xff00ff00),
       SnackBarVibe.warning: const Color(0xFFF46B22),
     };
 
     final Map<SnackBarVibe, Color> backgroundColors = {
-      // For every vibe, border is mainColor with 20% opacity.
-      for (final vibe in SnackBarVibe.values)
-        vibe: mainColors[vibe]!.withValues(alpha: 0.05),
+      for (final vibe in SnackBarVibe.values) vibe: mainColors[vibe]!.withValues(alpha: 0.2)
     };
-
-    final Map<SnackBarVibe, Color> borderColors = {
-      // For every vibe, border is mainColor with 20% opacity.
-      for (final vibe in SnackBarVibe.values)
-        vibe: mainColors[vibe]!.withValues(alpha: 0.2),
-    };
-
-
 
     final Map<SnackBarVibe, IconData> icons = {
-      SnackBarVibe.none: Icons.remove_circle_outline,
-      SnackBarVibe.info: Icons.info,
-      SnackBarVibe.error: Icons.error_outline,
-      SnackBarVibe.neutral: Icons.info_outline,
-      SnackBarVibe.success: Icons.check_circle_outline,
-      SnackBarVibe.warning: Icons.warning_amber_outlined,
+      SnackBarVibe.none: Icons.info_outline_rounded,
+      SnackBarVibe.info: Icons.info_outline_rounded,
+      SnackBarVibe.error: Icons.error_outline_rounded,
+      SnackBarVibe.success: Icons.check_circle_outline_rounded,
+      SnackBarVibe.warning: Icons.warning_amber_rounded,
     };
 
     // Choose colors based on whether dark mode is active.
@@ -93,25 +80,49 @@ class CustomSnackBar {
     final Color bgColor = backgroundColors[vibe]!;
     final IconData iconData = icons[vibe]!;
 
+    final ThemeData themeData = Theme.of(context);
+    final bool isDarkMode = themeData.brightness == Brightness.dark;
+    final Color primaryColor = themeData.primaryColor;
+    final Color scaffoldBackgroundColor = themeData.scaffoldBackgroundColor;
+
     // Remove any currently visible SnackBar.
     ScaffoldMessenger.of(context).hideCurrentSnackBar();
 
     // Show the custom SnackBar.
     ScaffoldMessenger.of(context).showSnackBar(
       SnackBar(
+        showCloseIcon: showCloseIcon,
+        width: width,
+        padding: padding,
         behavior: SnackBarBehavior.floating, // Makes the SnackBar float above the UI.
         duration: duration ?? const Duration(milliseconds: 2000),
         margin: margin ?? const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-        backgroundColor: backgroundColor ?? bgColor,
+        backgroundColor: backgroundColor ??
+            (usePrimaryColor && vibe == SnackBarVibe.none
+                ? primaryColor.withValues(alpha: 0.3)
+                : vibe == SnackBarVibe.none
+                    ? (scaffoldBackgroundColor.withValues(alpha: 0.25))
+                    : bgColor),
         dismissDirection: dismissDirection ?? DismissDirection.vertical,
-        elevation: 0,
+        elevation: 64,
         shape: shape ??
             RoundedRectangleBorder(
               borderRadius: BorderRadius.circular(6),
-              side: BorderSide(color: borderColors[vibe]!, width: 1),
+              side: BorderSide(
+                  color: borderColor ??
+                      (backgroundColor?.withValues(alpha: 0.75) ??
+                          (usePrimaryColor && vibe == SnackBarVibe.none
+                              ? primaryColor.withValues(alpha: 0.75)
+                              : vibe == SnackBarVibe.none
+                                  ? (scaffoldBackgroundColor.withValues(alpha: 0.75))
+                                  : bgColor.withValues(alpha: 0.75))),
+                  width: 1),
             ),
         content: BackdropFilter(
-          filter: ImageFilter.blur(sigmaX: 4, sigmaY: 4),
+          filter: ImageFilter.blur(
+            sigmaX: 4,
+            sigmaY: 4,
+          ),
           child: contentWidget ??
               Row(
                 mainAxisSize: MainAxisSize.min,
@@ -122,16 +133,17 @@ class CustomSnackBar {
                       content,
                       style: textStyle ??
                           TextStyle(
-                            color: mainColor,
+                            color: vibe == SnackBarVibe.none ? (isDarkMode ? Colors.black : Colors.white) : mainColor,
                             fontSize: 16,
                           ),
                     ),
                   ),
                   const SizedBox(width: 8),
-                  Icon(
-                    iconData,
-                    color: mainColor,
-                  ),
+                  icon ??
+                      Icon(
+                        iconData,
+                        color: vibe == SnackBarVibe.none ? (isDarkMode ? Colors.black : Colors.white) : mainColor,
+                      ),
                 ],
               ),
         ),